Pushed LongLongDataType.cpp

+++ If-Else/LongLongDataType.cpp
@@ -2,8 +2,8 @@
 #include<climits>
 using namespace std;
 int main(){
-    long long x = LLONG_MAX;
+    int x = INT_MAX;
     cout << x << endl;
-    long long y = LLONG_MIN;
+    int y = INT_MIN;
     cout << y << endl;
 }
